+++ src/meta_analyze/plotting/__init__.py
@@ -2,6 +2,7 @@
 
 from .forest import forest_plot
 from .funnel import funnel_plot
+from .regression import bubble_plot
 from .subgroup_forest import subgroup_forest_plot
 
-__all__ = ["forest_plot", "funnel_plot", "subgroup_forest_plot"]
+__all__ = ["bubble_plot", "forest_plot", "funnel_plot", "subgroup_forest_plot"]